Create 2_SQL_code.sql
Write some SQL code that generates one output with the following fields:
- Order_ID
- Status
- Country
- List of Categories (list separated by “|”) in case an order is assigned to more than 1
category
- Number of days between order and today (create custom function to calculate this
 
value)
- Order amount
- Country_order_value_index – this represents the position in terms of order value for
every country. For example, if this order has the highest order value for its Country, then it should show the value 1, if it’s the second highest then should show the value 2. As there are three countries, there should be three order ids with the value 1.